Refactor WebSockets in test to websocket-util module
